ZZFLEX引入

+++ TLChat/Modules/Chat/ChatViewController/ChatCommonViews/TLRecorderIndicatorView/TLRecorderIndicatorView.m
@@ -82,7 +82,7 @@
     _volume = volume;
     NSInteger picId = 10 * (volume < 0 ? 0 : (volume > 1.0 ? 1.0 : volume));
     picId = picId > 8 ? 8 : picId;
-    [self.volumeImageView setImage:[UIImage imageNamed:[NSString stringWithFormat:@"chat_record_signal_%ld", picId]]];
+    [self.volumeImageView setImage:[UIImage imageNamed:[NSString stringWithFormat:@"chat_record_signal_%ld", (long)picId]]];
 }
 
 #pragma mark - # Private Methods
